+++ thunder/tests/test_networks.py
@@ -2,13 +2,10 @@ import math
 import os
 from functools import partial
 import warnings
-from dataclasses import dataclass
-from contextlib import contextmanager
 from looseversion import LooseVersion
 
 import pytest
 import torch
-import torch.nn as nn
 from torch.testing import assert_close, make_tensor
 
 import thunder
@@ -22,6 +19,7 @@ from thunder.tests.framework import (
 )
 import thunder.tests.nanogpt_model as nanogpt_model
 import thunder.tests.hf_bart_self_attn as hf_bart_self_attn
+import thunder.tests.llama4_moe as llama4_moe
 
 #
 # nanoGPT tests
@@ -728,152 +726,12 @@ def test_hf_kvcache():
     assert_close(j_static_cache.value_cache, ref_static_cache.value_cache, rtol=1e-1, atol=1e-1)
 
 
-@dataclass(frozen=True)
-class Config:
-    name: str
-    hidden_size: int
-    intermediate_size: int
-    num_routed_experts: int
-    num_shared_experts: int
-    dtype: torch.dtype = torch.bfloat16
-    device: str = "cuda"
-
-
-small_config = Config(name="small", hidden_size=256, intermediate_size=512, num_routed_experts=8, num_shared_experts=1)
-llama4_config = Config(
+small_config = llama4_moe.Config(
+    name="small", hidden_size=256, intermediate_size=512, num_routed_experts=8, num_shared_experts=1
+)
+llama4_config = llama4_moe.Config(
     name="llama4", hidden_size=5120, intermediate_size=8192, num_routed_experts=128, num_shared_experts=1
 )
-
-
-class SwiGLU(nn.Module):
-    def __init__(self, hidden_size: int, intermediate_size: int, dtype: torch.dtype, device: str):
-        super().__init__()
-        self.gate_proj = nn.Linear(hidden_size, intermediate_size, bias=False, dtype=dtype, device=device)
-        self.up_proj = nn.Linear(hidden_size, intermediate_size, bias=False, dtype=dtype, device=device)
-        self.down_proj = nn.Linear(intermediate_size, hidden_size, bias=False, dtype=dtype, device=device)
-
-    def forward(self, hidden_states: torch.Tensor) -> torch.Tensor:
-        return self.down_proj(torch.nn.functional.silu(self.gate_proj(hidden_states)) * self.up_proj(hidden_states))
-
-
-def _group_sizes_from_offsets(offsets: torch.Tensor) -> list[int]:
-    group_sizes = []
-    prev = 0
-    for offset in offsets:
-        group_sizes.append(offset - prev)
-        prev = offset
-    return group_sizes
-
-
-if LooseVersion(torch.__version__) >= LooseVersion("2.8.0"):
-    # Required otherwise, there is a graph-break.
-    _grouped_mm = torch.compiler.allow_in_graph(torch._grouped_mm)
-
-
-# This function should be replaced with torch._grouped_mm.  However,
-# torch._grouped_mm is yet to be usable because it requires offsets being
-# multiples of 16.
-def grouped_mm(a: torch.Tensor, b: torch.Tensor, offsets: torch.Tensor) -> torch.Tensor:
-    if torch.compiler.is_compiling():
-        # NOTE: This path also works for `thunder.jit` as it has a lookaside for `torch.compiler.is_compiling`.
-        return _grouped_mm(a, b, offsets)
-
-    group_sizes = _group_sizes_from_offsets(offsets)
-    group_outs = []
-    for group_a, group_b in zip(a.split(group_sizes), b.unbind()):
-        group_outs.append(group_a @ group_b)
-    return torch.cat(group_outs)
-
-
-class GroupedLinear(nn.Module):
-    def __init__(self, groups: int, in_features: int, out_features: int, dtype: torch.dtype, device: str):
-        super().__init__()
-        self.weight = nn.Parameter(torch.empty(groups, in_features, out_features, dtype=dtype, device=device))
-        # Initialize the weight in the same way as nn.Linear
-        nn.init.kaiming_uniform_(self.weight, a=math.sqrt(5))
-
-    def forward(self, hidden_states: torch.Tensor, offsets: torch.Tensor) -> torch.Tensor:
-        return grouped_mm(hidden_states, self.weight, offsets)
-
-
-class GroupedSwiGLU(nn.Module):
-    def __init__(self, groups: int, hidden_size: int, intermediate_size: int, dtype: torch.dtype, device: str):
-        super().__init__()
-        self.gate_proj = GroupedLinear(groups, hidden_size, intermediate_size, dtype, device)
-        self.up_proj = GroupedLinear(groups, hidden_size, intermediate_size, dtype, device)
-        self.down_proj = GroupedLinear(groups, intermediate_size, hidden_size, dtype, device)
-
-    def forward(self, hidden_states: torch.Tensor, offsets: torch.Tensor) -> torch.Tensor:
-        return self.down_proj(
-            torch.nn.functional.silu(self.gate_proj(hidden_states, offsets)) * self.up_proj(hidden_states, offsets),
-            offsets,
-        )
-
-
-class Llama4MoE(nn.Module):
-    def __init__(self, config: Config):
-        super().__init__()
-        self.config = config
-        self.gate = nn.Linear(
-            config.hidden_size, config.num_routed_experts, bias=False, dtype=config.dtype, device=config.device
-        )
-        self.shared_experts = SwiGLU(
-            config.hidden_size, config.intermediate_size * config.num_shared_experts, config.dtype, config.device
-        )
-        self.routed_experts = GroupedSwiGLU(
-            config.num_routed_experts, config.hidden_size, config.intermediate_size, config.dtype, config.device
-        )
-
-    def run_routed_experts(self, hidden_states: torch.Tensor) -> torch.Tensor:
-        batch_size, seq_len, _ = hidden_states.size()
-        hidden_states = hidden_states.view(-1, hidden_states.size(-1))  # [s, h]
-
-        router_logits = self.gate(hidden_states)  # [s, n]
-        topk_weight, topk_ids = router_logits.topk(1)  # [s, 1]
-        router_scores = topk_weight.sigmoid()  # [s, 1]
-        hidden_states = hidden_states * router_scores  # [s, h]
-
-        counts = torch.zeros(
-            topk_ids.size(0),
-            self.config.num_routed_experts,
-            device=topk_ids.device,
-            dtype=torch.int32,
-        )  # [s, n]
-        counts = counts.scatter(1, topk_ids, 1)  # [s, n]
-        tokens_per_expert = counts.sum(0)  # [n]
-
-        token_ids_sorted_by_expert_id = topk_ids.view(-1).argsort()  # [s]
-        tokens_sorted_by_expert_id = hidden_states[token_ids_sorted_by_expert_id]  # [s, h]
-
-        # Without `torch.int32`, we see `RuntimeError: Offsets tensor must be integer (int32) tensor, but got torch.int64.`
-        # from PyTorch when calling _grouped_mm.
-        offsets = torch.cumsum(tokens_per_expert, 0, dtype=torch.int32)  # [n]
-        outs_sorted_by_expert_id = self.routed_experts(tokens_sorted_by_expert_id, offsets)  # [s, h]
-
-        token_ids_sorted_by_expert_inverse_id = torch.argsort(token_ids_sorted_by_expert_id)
-        outs_sorted_by_token_id = outs_sorted_by_expert_id[token_ids_sorted_by_expert_inverse_id]
-
-        return outs_sorted_by_token_id
-
-    def forward(self, hidden_states: torch.Tensor) -> torch.Tensor:
-        return self.shared_experts(hidden_states) + self.run_routed_experts(hidden_states)
-
-
-@contextmanager
-def default_tensor_type(dtype=torch.float32, device="cpu"):
-    # Save
-    prev_dtype = torch.get_default_dtype()
-    prev_device = torch.get_default_device()
-
-    # Set
-    torch.set_default_dtype(dtype)
-    torch.set_default_device(device)
-
-    yield
-
-    # Restore
-    torch.set_default_dtype(prev_dtype)
-    torch.set_default_device(prev_device)
 
 
 @requiresCUDA
@@ -889,7 +747,7 @@ def test_llama4_moe(jit_fn, config):
 
     @requiresDeviceMemory(required_memory_bytes=required_memory)
     def _test():
-        model = Llama4MoE(config)
+        model = llama4_moe.Llama4MoE(config)
 
         # Without this, `thunderfx` falls back to `inductor` for `_grouped_mm`
         # as it doesn't have a grad-rule for the same.
